+++ Injir/Model/Resources/Assets.xcassets/AccentColor.colorset/Contents.json
