+++ Injir/Model/Resources/Assets.xcassets/backArrow.imageset/solar_alt-arrow-left-line-duotone.pdf
