+++ Injir/Model/Resources/Assets.xcassets/flagsButtons/russia.imageset/Rectangle 20.pdf
