+++ Injir/Model/Resources/Assets.xcassets/flagsButtons/tajikistan.imageset/Contents.json
