+++ Injir/Model/Resources/Assets.xcassets/flagsButtons/tajikistan.imageset/Rectangle 22.png
