+++ Injir/Model/Resources/Assets.xcassets/flagsButtons/turkmenistan.imageset/Contents.json
