+++ Injir/Model/Resources/Assets.xcassets/flagsButtons/turkmenistan.imageset/Rectangle 23.png
