+++ Injir/Model/Resources/Assets.xcassets/flagsButtons/uzbekistan.imageset/Contents.json
