+++ Injir/Model/Resources/Assets.xcassets/flagsButtons/uzbekistan.imageset/Rectangle 21.png
