+++ Injir/Model/Resources/Assets.xcassets/logo.imageset/Contents.json
